pven/requirements.txt
environ
python-dotenv
mysqlclient==2.1.1
djangorestframework==3.14.0

